Move Ctx and TyCtx to a separate ctx module
